set chakra UI for FRONTEND and install all dependencies test chakraUI working or not

+++ backend/src/controllers/product.controller.js
@@ -29,6 +29,11 @@ export const createProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid product ID" });
+  }
 
   try {
     await Product.findByIdAndDelete(id);
